Added user velidation to all method, added logging for errors

+++ Server/Server.API/Controllers/UserController.cs
@@ -2,6 +2,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Server.API.DTO;
 using AutoMapper;
+using System.Security.Claims;
+using Azure.Core;
 
 namespace Server.API.Controllers;
 
@@ -25,14 +27,24 @@ public class UserController : Controller
     public async Task<IActionResult> UpdateUser(UserDTO request)
     {
         _logger.LogInformation($"Processing route: {Request.Path.Value}");
+
+        var userId = Convert.ToInt32(HttpContext.User.FindFirstValue("Id"));
+        if (userId != request.Id)
+        {
+            _logger.LogError($"{Request.Path.Value}: 403 Forbidden");
+            return Forbid();
+        }
+
         var user = await _mediator.Send(new GetUserByIdRequest(request.Id));
         if (user is null)
         {
+            _logger.LogError($"{Request.Path.Value}: 404 Not Found");
             return NotFound();
         }
         user.Name = request.Name;
         await _mediator.Send(new UpdateUserRequest(user));
         var userDto = _mapper.Map<UserDTO>(user);
+        _logger.LogInformation($"{Request.Path.Value}: 200 OK");
         return Ok(userDto);
     }
 
@@ -40,17 +52,28 @@ public class UserController : Controller
     public async Task<IActionResult> UpdateUserPassword(ChangePasswordDTO request)
     {
         _logger.LogInformation($"Processing route: {Request.Path.Value}");
+
+        var userId = Convert.ToInt32(HttpContext.User.FindFirstValue("Id"));
+        if (userId != request.Id)
+        {
+            _logger.LogError($"{Request.Path.Value}: 403 Forbidden");
+            return Forbid();
+        }
+
         var user = await _mediator.Send(new GetUserByIdRequest(request.Id));
         if (user is null)
         {
+            _logger.LogError($"{Request.Path.Value}: 404 Not Found");
             return NotFound();
         }
         if (user.Password != request.OldPassword)
         {
+            _logger.LogError($"{Request.Path.Value}: 400 Bad Request");
             return BadRequest();
         }
         user.Password = request.NewPassword;
         await _mediator.Send(new UpdateUserRequest(user));
+        _logger.LogInformation($"{Request.Path.Value}: 200 OK");
         return Ok();
     }
 
@@ -58,7 +81,16 @@ public class UserController : Controller
     public async Task<IActionResult> DeleteUser(int id)
     {
         _logger.LogInformation($"Processing route: {Request.Path.Value}");
+
+        var userId = Convert.ToInt32(HttpContext.User.FindFirstValue("Id"));
+        if (userId != id)
+        {
+            _logger.LogError($"{Request.Path.Value}: 403 Forbidden");
+            return Forbid();
+        }
+
         await _mediator.Send(new DeleteUserRequest(id));
+        _logger.LogInformation($"{Request.Path.Value}: 200 OK");
         return Ok();
     }
 
@@ -66,19 +98,28 @@ public class UserController : Controller
     public async Task<IActionResult> GetUserChats(int id)
     {
         _logger.LogInformation($"Processing route: {Request.Path.Value}");
+
+        var userId = Convert.ToInt32(HttpContext.User.FindFirstValue("Id"));
+        if (userId != id)
+        {
+            _logger.LogError($"{Request.Path.Value}: 403 Forbidden");
+            return Forbid();
+        }
+
         var chats = await _mediator.Send(new GetUserChatsRequest(id));
         var chatsDto = _mapper.Map<IEnumerable<ChatDTO>>(chats);
+        _logger.LogInformation($"{Request.Path.Value}: 200 OK");
         return Ok(chatsDto);
     }
 
-    [HttpGet("chatMessages/userId={userId:int}&chatId={chatId:int}")]
-    public async Task<IActionResult> GetUserChatMessages(int userId, int chatId)
-    {
-        _logger.LogInformation($"Processing route: {Request.Path.Value}");
-        var messages = await _mediator.Send(new GetUserChatMessagesRequest(userId, chatId));
-        var messagesDto = _mapper.Map<IEnumerable<MessageDTO>>(messages);
-        return Ok(messagesDto);
-    }
+    //[HttpGet("chatMessages/userId={userId:int}&chatId={chatId:int}")]
+    //public async Task<IActionResult> GetUserChatMessages(int userId, int chatId)
+    //{
+    //    _logger.LogInformation($"Processing route: {Request.Path.Value}");
+    //    var messages = await _mediator.Send(new GetUserChatMessagesRequest(userId, chatId));
+    //    var messagesDto = _mapper.Map<IEnumerable<MessageDTO>>(messages);
+    //    return Ok(messagesDto);
+    //}
 
     [HttpGet("allUsers")]
     public async Task<IActionResult> GetAllUsersRequest()
@@ -86,6 +127,7 @@ public class UserController : Controller
         _logger.LogInformation($"Processing route: {Request.Path.Value}");
         var users = await _mediator.Send(new GetAllUsersRequest());
         var usersDto = _mapper.Map<IEnumerable<UserDTO>>(users);
+        _logger.LogInformation($"{Request.Path.Value}: 200 OK");
         return Ok(usersDto);
     }
 }
